cleanup common components
