+++ src/app/views/common/copy-button/CopyButton.tsx
@@ -1,54 +1,47 @@
-import { IButton, IconButton, IIconProps, PrimaryButton } from '@fluentui/react';
-import { createRef, useState } from 'react';
+import { Button } from '@fluentui/react-components';
+import { CheckmarkRegular, CopyRegular } from '@fluentui/react-icons';
+import { useState } from 'react';
 import { translateMessage } from '../../../utils/translate-messages';
 
 interface ICopyButtonProps {
-  style?: any;
-  handleOnClick: Function;
-  className?: any;
+  handleOnClick: () => void;
   isIconButton: boolean;
 }
 
-export default function CopyButton(props:ICopyButtonProps) {
+export default function CopyButton(props: ICopyButtonProps) {
   const [copied, setCopied] = useState(false);
 
-  const copyIcon: IIconProps = {
-    iconName: !copied ? 'Copy' : 'CheckMark'
-  };
-
-  const copyLabel: string = !copied ? translateMessage('Copy') : translateMessage('Copied');
-
-  const copyRef = createRef<IButton>();
-  const onSetFocus = () => copyRef.current!.focus();
+  const CopyIcon = !copied ? CopyRegular : CheckmarkRegular;
+  const copyLabel: string = !copied
+    ? translateMessage('Copy')
+    : translateMessage('Copied');
 
   const handleCopyClick = async () => {
-    props.handleOnClick(props);
+    props.handleOnClick();
     setCopied(true);
     handleTimeout();
-    onSetFocus();
   };
 
   const handleTimeout = () => {
-    const timer = setTimeout(() => { setCopied(false) }, 3000); // 3 seconds
+    const timer = setTimeout(() => {
+      setCopied(false);
+    }, 3000); // 3 seconds
     return () => clearTimeout(timer);
-  }
+  };
 
   return (
     <>
-      {props.isIconButton ?
-        <IconButton
-          toggle
+      {props.isIconButton ? (
+        <Button
+          appearance='transparent'
           onClick={handleCopyClick}
-          iconProps={copyIcon}
-          title={copyLabel}
-          ariaLabel={copyLabel}
-          style={props.style}
-          className={props.className}
-          componentRef={copyRef}
+          icon={<CopyIcon />}
         />
-        :
-        <PrimaryButton onClick={handleCopyClick} text={copyLabel} componentRef={copyRef}/>
-      }
+      ) : (
+        <Button appearance='transparent' onClick={handleCopyClick}>
+          {copyLabel}
+        </Button>
+      )}
     </>
-  )
+  );
 }
