+++ src/app/views/common/copy-button/index.ts
@@ -1,3 +1,2 @@
 import CopyButton from './CopyButton';
-import CopyButtonV9 from './CopyButtonV9';
-export { CopyButton, CopyButtonV9 };
+export { CopyButton };
